requirements.txt
python-dotenv
djangorestframework
django-cors-headers
Django
pytz
sqlparse
